refactor: move lexical string primitives
