+++ src/language/dml/parse-update.ts
@@ -8,7 +8,7 @@ import type {
   TakeUntilTopLevelKeyword,
   Trim,
   Unquote,
-} from '../../string.js';
+} from '../lexical/string.js';
 import type { PredicateIR } from '../ir/predicate.js';
 import type { UpdateQueryIR } from '../ir/query.js';
 import type { AssignmentIR, WriteTargetIR } from '../ir/write.js';
